+++ Utilities/DumpData/mod_info.js
@@ -3,10 +3,17 @@
    Lang : "*",
    Duration : "temp",
    Name : "Dump Data",
-   Description : "Export various data to xml/csv when you start or load a geoscape game.  Note that Dump_Others requires Scripting Library.\n\nData are saved in this mod's folder by default - click the path below to open the folder.",
+   Description : "
+Export various data to xml/csv when you start or load a geoscape game.
+   
+Note that Dump_Others config requires JavaScript Runtime, but the other data can be exported standalone.
+
+Data are saved in this mod's folder by default.
+",
    Url : {
       "GitHub" : "https://github.com/Sheep-y/PhoenixPt-Mods/",
       "Nexus"  : "https://nexusmods.com/phoenixpoint/mods/50/",
+      "JavaScript Runtime" : "https://www.nexusmods.com/phoenixpoint/mods/49",
    },
    ConfigType : "Sheepy.PhoenixPt.DumpData.ModConfig",
 })